split blockpage tests
